added an option to switch between following posts and all posts

+++ client/src/components/popover/BasicPopover.js
@@ -8,10 +8,12 @@ import { deletePost } from "../../utils/api";
 import { AuthContext } from "../../context/AuthContext";
 import { useHistory } from "react-router-dom";
 
-export function BasicPopover({ postId }) {
+export function BasicPopover({ postId, userId }) {
   const [anchorEl, setAnchorEl] = React.useState(null);
   const currentUser = React.useContext(AuthContext).user;
   const history = useHistory();
+
+  console.log(userId, currentUser._id);
 
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -23,16 +25,14 @@ export function BasicPopover({ postId }) {
 
   const handleCopy = () => {
     //add actual link
-    navigator.clipboard.writeText("heba");
+    navigator.clipboard.writeText(`posts/${postId}`);
   };
 
   const handleDelete = () => {
     try {
-      console.log("-----------");
-      console.log(postId);
-      console.log(currentUser._id);
-
-      deletePost(postId, currentUser._id).then(history.go());
+      if (window.confirm("Delete post?")) {
+        deletePost(postId, currentUser._id).then(history.go());
+      }
     } catch (error) {
       console.log(error);
     }
@@ -62,10 +62,14 @@ export function BasicPopover({ postId }) {
           <div className="popoverItem" onClick={handleCopy}>
             copy link
           </div>
-          <hr className="popoverHr" />
-          <div className="popoverItem" onClick={handleDelete}>
-            delete
-          </div>
+          {userId === currentUser._id ? (
+            <>
+              <hr className="popoverHr" />
+              <div className="popoverItem" onClick={handleDelete}>
+                delete
+              </div>
+            </>
+          ) : null}
         </Typography>
       </Popover>
     </div>
